# Source/Backend/Golang/cart/internal/delivery/http/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Сервис корзины",
        "title": "Cart Service",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8089",
    "basePath": "/v1",
    "paths": {
        "/cart": {
            "get": {
                "security": [
                    {
                        "AccountId": []
                    },
                    {
                        "AccountType": []
                    }
                ],
                "description": "Получение товаров из корзины покупателя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cart"
                ],
                "summary": "Получение товаров из корзины",
                "operationId": "get-cart",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/viewmodel.Product"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responses.StatusResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.StatusResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "AccountId": []
                    },
                    {
                        "AccountType": []
                    }
                ],
                "description": "Очистка корзины покупателя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cart"
                ],
                "summary": "Очистка корзины",
                "operationId": "clear-cart",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responses.StatusResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.StatusResponse"
                        }
                    }
                }
            }
        },
        "/cart/checkout": {
            "post": {
                "security": [
                    {
                        "AccountId": []
                    },
                    {
                        "AccountType": []
                    }
                ],
                "description": "Оформление товаров из корзины покупателя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cart"
                ],
                "summary": "Оформление товаров из корзины",
                "operationId": "cart-checkout",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responses.StatusResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.StatusResponse"
                        }
                    }
                }
            }
        },
        "/cart/products": {
            "post": {
                "security": [
                    {
                        "AccountId": []
                    },
                    {
                        "AccountType": []
                    }
                ],
                "description": "Добавление товара в корзину покупателя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cart"
                ],
                "summary": "Добавление товара в корзину",
                "operationId": "cart-add-product",
                "parameters": [
                    {
                        "description": "Товар",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/viewmodel.AddProduct"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.StatusResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responses.StatusResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.StatusResponse"
                        }
                    }
                }
            }
        },
        "/cart/products/{id}": {
            "delete": {
                "security": [
                    {
                        "AccountId": []
                    },
                    {
                        "AccountType": []
                    }
                ],
                "description": "Удаление товара из корзины покупателя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cart"
                ],
                "summary": "Удаление товара из корзины",
                "operationId": "cart-delete-product",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Идентификатор товара",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.StatusResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responses.StatusResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.StatusResponse"
                        }
                    }
                }
            }
        },
        "/cart/products/{id}/count": {
            "put": {
                "security": [
                    {
                        "AccountId": []
                    },
                    {
                        "AccountType": []
                    }
                ],
                "description": "Обновить количество товара в корзине покупателя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cart"
                ],
                "summary": "Обновить количество товара в корзине",
                "operationId": "cart-change-product-count",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Идентификатор товара",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Количество",
                        "name": "count",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/viewmodel.AddCount"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.StatusResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responses.StatusResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.StatusResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "responses.StatusResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "viewmodel.AddCount": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                }
            }
        },
        "viewmodel.AddProduct": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "viewmodel.Product": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "AccountId": {
            "type": "apiKey",
            "name": "X-Account-Id",
            "in": "header"
        },
        "AccountType": {
            "type": "apiKey",
            "name": "X-Account-Type",
            "in": "header"
        }
    }
}